+++ tosho/src/impl/amap/purchases.rs
@@ -72,6 +72,8 @@ pub(crate) async fn amap_purchase(
                             continue;
                         }
 
+                        super::common::save_session_config(&client, &account);
+
                         // Sleep for 500ms to avoid being too fast
                         // and made the claiming failed
                         sleep(Duration::from_millis(500)).await;